flock.c: Ignore bogus samples when nbhd population is low
During sample generation, if the neighborhood around a given
boid is sparsely populated, the sample.indices array can
have unpopulated slots that remain set to zero. This causes
the boids to concentrate around the corners of the simulation,
and generally break it.

This patch initializes the array to -1, instead of zero, and
any slots that remain unfilled are ignored.